movie-recommender/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.2.6
scikit-learn==1.7.1
scipy==1.15.3
sqlalchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
joblib==1.5.1
requests==2.32.4
python-multipart==0.0.20
